+++ NEWS.md
@@ -12,6 +12,7 @@
    
 ## Minor changes
    
+- the results of `t_test()`, `wilcox_test()`, `dunn_test()` and `sign_test()` are now an object of class `rstatix_test` that has an attribute named **args** for holding the test arguments.
 - The results of `cohens_d()` is now a data frame containing the Cohen's d and the magnitude.
 
 ## Bug fixes